changing the slit math and algorithm to make it work
The math for the slit scan works for y. Now to adjust x accordingly.

adjusted the growth parameter

The basics of the real slitscan is almost there

creating the final test to get it right.

The math is almost there.

It's kinda working in SSTest3, except not a linear expansion.

SSTest4 is perfect. Will copy this into SSTest later and delete the
rest.

cleanup of all the intermediate tests